+++ CHANGELOG.md
@@ -7,8 +7,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## 4.0.0 - 2024-03-26
 ### Added
 - Added support for Craft 5
+
 ### Changed
 - Updated build tools and Node dependencies
+
+### Removed
+- Removed the following Grid Components and Snippets from Guide Editor
+  - Grid
+  - Grid – Custom Columns
+  - Element Search
 
 
 ## 3.2.2 - 2023-11-13